Move existing Quack component into Chat component in order to render other components with routing
